+++ sdmx-dl-provider-ri/src/test/java/internal/sdmxdl/provider/ri/web/Sdmx21RestQueriesTest.java
@@ -19,10 +19,7 @@ package internal.sdmxdl.provider.ri.web;
 import nbbrd.io.text.Parser;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import sdmxdl.DataQuery;
-import sdmxdl.DataStructureRef;
-import sdmxdl.DataflowRef;
-import sdmxdl.Key;
+import sdmxdl.*;
 import sdmxdl.provider.DataRef;
 
 import java.net.URL;
@@ -47,9 +44,9 @@ public class Sdmx21RestQueriesTest {
             assertThatNullPointerException().isThrownBy(() -> x.getFlowQuery(base, null));
             assertThatNullPointerException().isThrownBy(() -> x.getStructureQuery(null, specificStruct));
             assertThatNullPointerException().isThrownBy(() -> x.getStructureQuery(base, null));
-            assertThatNullPointerException().isThrownBy(() -> x.getDataQuery(null, DataRef.of(specificFlow, DataQuery.of(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef));
+            assertThatNullPointerException().isThrownBy(() -> x.getDataQuery(null, DataRef.of(specificFlow, newDataQuery(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef));
             assertThatNullPointerException().isThrownBy(() -> x.getDataQuery(base, null, ignoreDsdRef));
-            assertThatNullPointerException().isThrownBy(() -> x.getDataQuery(base, DataRef.of(specificFlow, DataQuery.of(Key.ALL, SERIES_KEYS_ONLY)), null));
+            assertThatNullPointerException().isThrownBy(() -> x.getDataQuery(base, DataRef.of(specificFlow, newDataQuery(Key.ALL, SERIES_KEYS_ONLY)), null));
         }
     }
 
@@ -97,23 +94,23 @@ public class Sdmx21RestQueriesTest {
     public void testGetDataQuery() {
         Assertions.assertThat(Sdmx21RestQueries.builder().build())
                 .satisfies(x -> {
-                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, DataQuery.of(Key.ALL, FULL)), ignoreDsdRef))
+                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, newDataQuery(Key.ALL, FULL)), ignoreDsdRef))
                             .hasToString("http://base/data/ECB%2CEXR%2C1.0/all/all");
 
-                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, DataQuery.of(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef))
+                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, newDataQuery(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef))
                             .hasToString("http://base/data/ECB%2CEXR%2C1.0/all/all?detail=serieskeysonly");
 
-                    assertThat(x.getDataQuery(base, DataRef.of(genericFlow, DataQuery.of(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef))
+                    assertThat(x.getDataQuery(base, DataRef.of(genericFlow, newDataQuery(Key.ALL, SERIES_KEYS_ONLY)), ignoreDsdRef))
                             .hasToString("http://base/data/all%2CEXR%2Clatest/all/all?detail=serieskeysonly");
 
-                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, DataQuery.of(key, SERIES_KEYS_ONLY)), ignoreDsdRef))
+                    assertThat(x.getDataQuery(base, DataRef.of(specificFlow, newDataQuery(key, SERIES_KEYS_ONLY)), ignoreDsdRef))
                             .hasToString("http://base/data/ECB%2CEXR%2C1.0/D.NOK.EUR.SP00.A/all?detail=serieskeysonly");
 
-                    assertThat(x.getDataQuery(base, DataRef.of(genericFlow, DataQuery.of(key, SERIES_KEYS_ONLY)), ignoreDsdRef))
+                    assertThat(x.getDataQuery(base, DataRef.of(genericFlow, newDataQuery(key, SERIES_KEYS_ONLY)), ignoreDsdRef))
                             .hasToString("http://base/data/all%2CEXR%2Clatest/D.NOK.EUR.SP00.A/all?detail=serieskeysonly");
                 });
 
-        Assertions.assertThat(Sdmx21RestQueries.builder().trailingSlashRequired(true).build().getDataQuery(base, DataRef.of(specificFlow, DataQuery.of(Key.ALL, FULL)), ignoreDsdRef))
+        Assertions.assertThat(Sdmx21RestQueries.builder().trailingSlashRequired(true).build().getDataQuery(base, DataRef.of(specificFlow, newDataQuery(Key.ALL, FULL)), ignoreDsdRef))
                 .hasToString("http://base/data/ECB%2CEXR%2C1.0/all/all/");
     }
 
@@ -124,4 +121,8 @@ public class Sdmx21RestQueriesTest {
     private final DataStructureRef genericStruct = DataStructureRef.of(null, "EXR", null);
     private final Key key = Key.parse("D.NOK.EUR.SP00.A");
     private final DataStructureRef ignoreDsdRef = DataStructureRef.parse("abc");
+
+    private static DataQuery newDataQuery(Key key, DataDetail detail) {
+        return DataQuery.builder().key(key).detail(detail).build();
+    }
 }