+++ sdmx-dl-provider-util/src/main/java/sdmxdl/provider/file/CachedFileClient.java
@@ -101,7 +101,7 @@ public final class CachedFileClient implements FileClient {
     }
 
     private DataSet copyAllNoData(FileInfo entry, DataflowRef flowRef) throws IOException {
-        DataRef ref = DataRef.of(flowRef, DataQuery.of(Key.ALL, DataDetail.NO_DATA));
+        DataRef ref = DataRef.of(flowRef, DataQuery.builder().key(Key.ALL).detail(DataDetail.NO_DATA).build());
         try (Stream<Series> stream = delegate.loadData(entry, ref)) {
             return stream.collect(toDataSet(ref.getFlowRef(), ref.getQuery()));
         }